Some bugs fixed.

+++ src/heavyweight/nodes/Diffuse.scala
@@ -8,7 +8,7 @@ case class Diffuse(override val inputs: Array[Container], override val outputs: 
 
   override def scatter(mesh: Mesh, world: World, triangleIndex: Int,
                        ray: Ray, hitPoint: Vector3D, coordinates: Vector3D, shadersLeft: Int,
-                       color: Color, roughness: Double, rays: Int,
+                       color: Color, roughness: Float, rays: Int,
                        normalMap: Vector3D, flatNormal: Vector3D,
                        integral: LampIlluminationOutput): Array[Ray] = RayDistributor.getRandomRays(
     flatNormal,
