+++ src/heavyweight/nodes/VolumeScatter.scala
@@ -13,8 +13,8 @@ case class VolumeScatter(override val inputs: Array[Container], override val out
   */
   override def doThings(mesh: Mesh, world: World, triangleIndex: Int, ray: Ray, hitPoint: Vector3D, coordinates: Vector3D, backColor: Color, shadersLeft: Int): Unit = {
     val volumeColor: Color = inputs(0).content.asInstanceOf[Color]
-    val volumeDensity: Double = inputs(1).content.asInstanceOf[lightweight.nodes.Numeric].value
-    val volumeAnisotropic: Double = inputs(2).content.asInstanceOf[lightweight.nodes.Numeric].value
+    val volumeDensity: Float = inputs(1).content.asInstanceOf[lightweight.nodes.Numeric].value
+    val volumeAnisotropic: Float = inputs(2).content.asInstanceOf[lightweight.nodes.Numeric].value
     val volumeNormal: Vector3D = inputs(3).content.asInstanceOf[Vector3D]
     val integral: LampIlluminationOutput = if (inputs(4) != null) inputs(4).content.asInstanceOf[LampIlluminationOutput] else LampIlluminationOutput(Color(0, 0, 0), 0)
     var color = Color(0, 0, 0)
