+++ src/lightweight/World.scala
@@ -2,4 +2,4 @@ package lightweight
 
 import lightweight.nodes.{SurfaceOutput, VolumeOutput}
 
-case class World(skySurface: SurfaceOutput, skyVolume: VolumeOutput, worldIOR: Double)
+case class World(skySurface: SurfaceOutput, skyVolume: VolumeOutput, worldIOR: Float)
